fix: change market_cap_rank to BIGINT and wire missing tests into CI
Python integers map to LongType (64-bit) in Spark; Delta rejects a merge
of LongType into INT (IntegerType/32-bit) with DELTA_MERGE_INCOMPATIBLE_DATATYPE.
Changing the Bronze table DDL to BIGINT aligns schema with the ingestion job output.

Also adds test_bronze_market_table_migration_job and test_silver_market_pipeline_job
to the CI contract job — both were committed but not yet wired into the pipeline.

+++ databricks/bronze_market_table_migration.sql
@@ -16,7 +16,7 @@ CREATE OR REPLACE TABLE cgadev.market_bronze.bronze_market_snapshots (
   price_usd DECIMAL(38, 8),
   volume_24h_usd DECIMAL(38, 8),
   circulating_supply DECIMAL(38, 8),
-  market_cap_rank INT,
+  market_cap_rank BIGINT,
   payload_version STRING
 );
